add jest testing libraries for typescript
