+++ package.json
@@ -21,13 +21,15 @@
   "homepage": "https://github.com/ram1117/capstone-2-fansite#readme",
   "devDependencies": {
     "@babel/plugin-transform-modules-commonjs": "^7.20.11",
+    "@types/jest": "^29.5.3",
     "css-loader": "^6.7.3",
     "html-loader": "^4.2.0",
     "html-webpack-plugin": "^5.5.0",
-    "jest": "^29.4.1",
+    "jest": "^29.6.2",
     "jest-environment-jsdom": "^29.4.1",
     "jsdom": "^21.1.0",
     "style-loader": "^3.3.1",
+    "ts-jest": "^29.1.1",
     "ts-loader": "^9.4.4",
     "typescript": "^5.1.6",
     "webpack": "^5.75.0",
